Moved playlist management endpoints to channels module
